requirements.txt
absl-py==1.0.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.5
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.4.0
autopep8==2.0.0
azure-core==1.26.3
azure-storage-blob==12.14.1
Babel==2.10.1
backcall==0.2.0
bleach==5.0.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==39.0.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.8.2
Flask==2.2.2
idna==3.3
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
json5==0.9.8
JsonForm==0.0.2
jsonschema==4.5.1
JsonSir==0.0.2
libclang==14.0.1
make==0.1.6.post2
Markdown==3.3.6
matplotlib-inline==0.1.3
numpy==1.22.3
oauthlib==3.2.0
pandas==1.5.0
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
virtualenv==20.17.1
